Update dependencies so that a new enough mochilo is used
In the gemspec, limit the mochilo version to something no older than
1.3 (brianmario/mochilo#20) for the 1.x series, and no older than 2.1
(brianmario/mochilo#19) for the 2.x series.

In the main Gemfile and in the v2 CI gemfile, bundle from github.com
because there hasn't been a gem cut yet.

In the v1 CI gemfile, bundle from the PR branch.