decode: define BERT module instead of trying to look it up
We want to make sure we are dealing with a module as we expect. Looking up the
object seems to result in the runtime re-using the returned `VALUE` leading to
us calling `supports?` on arbitrary objects when trying to decode.

The existing pattern for re-opening modules is to use `rb_define_module` so
let's do this.