decode: do not keep a reference to classes across calls
We do not want to keep the `Bert::Tuple` and `Mochilo` references and re-use
those values across calls. This has never been something we should do, but with
the advent of the compacting GC this becomes even more important.

Look up these values if we need them to be used only inside a single call into
the C code.